Add support for an Ugando osm file
In addition to the filename.shp input, a filename.osm can be given and a
filename_out.osm will be output.
Note that with an osm file the output will cleaner, none of the shapefile
specific hack will be done :
- coordinates will not be rounded to 5 decimals
- no inner ring will not be droped
- admin_level, name, old_name tag will be preserved

Fix simplification of closed line, if too small line will be removed.